feat(metrics): migrate the dashboard to MinIO Metrics V3
Rewrite every widget PromQL from V2 to V3 names: max()/min() dedup for
cluster gauges that each node exports identically, explicit zero-guards
for the server's zero-value skip (offline counts, traffic totals), the
api->name label rename, and seconds-based durations. Replace the V2-only
heal/scan activity widgets with Erasure Health (write-quorum baseline
guard) and Usage Data Age, drop unreferenced widgets 51/61/62, fix the
duplicated internode/syscall targets, and remove the unused Prometheus
label-values prefetch that stalled every widget request up to a second.

Zero-state semantics were then hardened against an adversarial review:
baseline the capacity free/used targets on the always-present total so a
full cluster reads 0 free instead of vanishing; guard Online Drives
against the all-offline zero-skip; switch bucket/object-count guards to
the usage group's own freshness gauge so pre-scan clusters read no-data
instead of a fabricated 0; render empty single-rep results as "—" rather
than 0; stop fabricating seven zero bins for an empty size distribution;
keep fractional rate() values visible (parseFloat axis domain, 2-decimal
CPU formatter); clamp sub-second Usage Data Age to "1 second".

Adds a regression suite pinning widget queries to the V3 catalog, the
per-widget guard taxonomy, a V2->V3 mapping document, and the zh
dictionary entries for the new cards. Ref pgsty/silo-console#7 #8.

+++ web-app/src/screens/Console/Dashboard/Prometheus/Widgets/LinearGraphWidget.tsx
@@ -155,7 +155,7 @@ const LinearGraphWidget = ({
               if (key === "name") {
                 continue;
               }
-              let val = parseInt(dp[key]);
+              let val = parseFloat(dp[key]);
 
               if (isNaN(val)) {
                 val = 0;
